TOOLS: move `CommandManager` to `tools-repl`
