+++ cli/src/main/java/module-info.java
@@ -12,7 +12,6 @@ module org.aya.cli {
   exports org.aya.cli;
   exports org.aya.cli.single;
   exports org.aya.cli.utils;
-  exports org.aya.cli.repl.command;
   exports org.aya.cli.library to com.google.gson;
   exports org.aya.cli.library.json to com.google.gson;
   exports org.aya.cli.repl;
