+++ tools-repl/src/main/java/org/aya/repl/Command.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2021 Yinsen (Tesla) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.repl.command;
+package org.aya.repl;
 
 import kala.collection.immutable.ImmutableSeq;
 import org.aya.pretty.doc.Doc;
