+++ tools-repl/src/main/java/org/aya/repl/CommandManager.java
@@ -1,12 +1,11 @@
 // Copyright (c) 2020-2021 Yinsen (Tesla) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.repl.command;
+package org.aya.repl;
 
 import kala.collection.immutable.ImmutableMap;
 import kala.collection.immutable.ImmutableSeq;
 import kala.control.Try;
 import kala.tuple.Tuple;
-import org.aya.repl.CommandArg;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
